table VO 수정
