+++ pet/src/main/java/com/pet/model/ReviewVO.java
@@ -3,17 +3,15 @@ package com.pet.model;
 public class ReviewVO {
 
 	private int num;
-	private String username;
-	private String pettype;
-	private double petweight;
-	private int petcount;
-	private String petinfo;
-	private String startdate;
-	private String enddate;
-	private String resdate;
-	private int price;
+	private String userid;
+	private String userpass;
+	private String img;
+	private String subject;
+	private String content;
 	private int resnum;
-	private int condition;
+	private int star;
+	private int rvnum;
+	private String wdate;
 	
 	
 	public int getNum() {
@@ -22,59 +20,35 @@ public class ReviewVO {
 	public void setNum(int num) {
 		this.num = num;
 	}
-	public String getUsername() {
-		return username;
+	public String getUserid() {
+		return userid;
 	}
-	public void setUsername(String username) {
-		this.username = username;
+	public void setUserid(String userid) {
+		this.userid = userid;
 	}
-	public String getPettype() {
-		return pettype;
+	public String getUserpass() {
+		return userpass;
 	}
-	public void setPettype(String pettype) {
-		this.pettype = pettype;
+	public void setUserpass(String userpass) {
+		this.userpass = userpass;
 	}
-	public double getPetweight() {
-		return petweight;
+	public String getImg() {
+		return img;
 	}
-	public void setPetweight(double petweight) {
-		this.petweight = petweight;
+	public void setImg(String img) {
+		this.img = img;
 	}
-	public int getPetcount() {
-		return petcount;
+	public String getSubject() {
+		return subject;
 	}
-	public void setPetcount(int petcount) {
-		this.petcount = petcount;
+	public void setSubject(String subject) {
+		this.subject = subject;
 	}
-	public String getPetinfo() {
-		return petinfo;
+	public String getContent() {
+		return content;
 	}
-	public void setPetinfo(String petinfo) {
-		this.petinfo = petinfo;
-	}
-	public String getStartdate() {
-		return startdate;
-	}
-	public void setStartdate(String startdate) {
-		this.startdate = startdate;
-	}
-	public String getEnddate() {
-		return enddate;
-	}
-	public void setEnddate(String enddate) {
-		this.enddate = enddate;
-	}
-	public String getResdate() {
-		return resdate;
-	}
-	public void setResdate(String resdate) {
-		this.resdate = resdate;
-	}
-	public int getPrice() {
-		return price;
-	}
-	public void setPrice(int price) {
-		this.price = price;
+	public void setContent(String content) {
+		this.content = content;
 	}
 	public int getResnum() {
 		return resnum;
@@ -82,19 +56,33 @@ public class ReviewVO {
 	public void setResnum(int resnum) {
 		this.resnum = resnum;
 	}
-	public int getCondition() {
-		return condition;
+	public int getStar() {
+		return star;
 	}
-	public void setCondition(int condition) {
-		this.condition = condition;
+	public void setStar(int star) {
+		this.star = star;
+	}
+	public int getRvnum() {
+		return rvnum;
+	}
+	public void setRvnum(int rvnum) {
+		this.rvnum = rvnum;
+	}
+	public String getWdate() {
+		return wdate;
+	}
+	public void setWdate(String wdate) {
+		this.wdate = wdate;
 	}
 	
 	
 	@Override
 	public String toString() {
-		return "ReviewVO [num=" + num + ", username=" + username + ", pettype=" + pettype + ", petweight=" + petweight
-				+ ", petcount=" + petcount + ", petinfo=" + petinfo + ", startdate=" + startdate + ", enddate="
-				+ enddate + ", resdate=" + resdate + ", price=" + price + ", resnum=" + resnum + ", condition="
-				+ condition + "]";
+		return "ReviewVO [num=" + num + ", userid=" + userid + ", userpass=" + userpass + ", img=" + img + ", subject="
+				+ subject + ", content=" + content + ", resnum=" + resnum + ", star=" + star + ", rvnum=" + rvnum
+				+ ", wdate=" + wdate + "]";
 	}
+	
+	
+	
 }
