+++ pet/src/main/java/com/pet/model/RoomVO.java
@@ -2,85 +2,56 @@ package com.pet.model;
 
 public class RoomVO {
 
-	private int num;
-	private String userid;
-	private String userpass;
-	private String img;
-	private String subject;
-	private String content;
-	private int resnum;
-	private int star;
-	private int rvnum;
-	private String wdate;
-	
-	
-	public int getNum() {
-		return num;
+	private String hotelname;
+	private String roomcode;
+	private String roomname;
+	private int roomprice;
+	private String roominfo;
+	private int petweight;
+	public String getHotelname() {
+		return hotelname;
 	}
-	public void setNum(int num) {
-		this.num = num;
+	public void setHotelname(String hotelname) {
+		this.hotelname = hotelname;
 	}
-	public String getUserid() {
-		return userid;
+	public String getRoomcode() {
+		return roomcode;
 	}
-	public void setUserid(String userid) {
-		this.userid = userid;
+	public void setRoomcode(String roomcode) {
+		this.roomcode = roomcode;
 	}
-	public String getUserpass() {
-		return userpass;
+	public String getRoomname() {
+		return roomname;
 	}
-	public void setUserpass(String userpass) {
-		this.userpass = userpass;
+	public void setRoomname(String roomname) {
+		this.roomname = roomname;
 	}
-	public String getImg() {
-		return img;
+	public int getRoomprice() {
+		return roomprice;
 	}
-	public void setImg(String img) {
-		this.img = img;
+	public void setRoomprice(int roomprice) {
+		this.roomprice = roomprice;
 	}
-	public String getSubject() {
-		return subject;
+	public String getRoominfo() {
+		return roominfo;
 	}
-	public void setSubject(String subject) {
-		this.subject = subject;
+	public void setRoominfo(String roominfo) {
+		this.roominfo = roominfo;
 	}
-	public String getContent() {
-		return content;
+	public int getPetweight() {
+		return petweight;
 	}
-	public void setContent(String content) {
-		this.content = content;
-	}
-	public int getResnum() {
-		return resnum;
-	}
-	public void setResnum(int resnum) {
-		this.resnum = resnum;
-	}
-	public int getStar() {
-		return star;
-	}
-	public void setStar(int star) {
-		this.star = star;
-	}
-	public int getRvnum() {
-		return rvnum;
-	}
-	public void setRvnum(int rvnum) {
-		this.rvnum = rvnum;
-	}
-	public String getWdate() {
-		return wdate;
-	}
-	public void setWdate(String wdate) {
-		this.wdate = wdate;
+	public void setPetweight(int petweight) {
+		this.petweight = petweight;
 	}
 	
 	
 	@Override
 	public String toString() {
-		return "RoomVO [num=" + num + ", userid=" + userid + ", userpass=" + userpass + ", img=" + img + ", subject="
-				+ subject + ", content=" + content + ", resnum=" + resnum + ", star=" + star + ", rvnum=" + rvnum
-				+ ", wdate=" + wdate + "]";
+		return "RoomVO [hotelname=" + hotelname + ", roomcode=" + roomcode + ", roomname=" + roomname + ", roomprice="
+				+ roomprice + ", roominfo=" + roominfo + ", petweight=" + petweight + "]";
 	}
+	
+	
 	
 }
